terminada el delete API

+++ api/controller/CommentApiController.php
@@ -5,6 +5,7 @@ require_once 'API/View/ApiView.php';
 require_once './model/CommentModel.php';
 require_once './controller/LoginController.php';
 
+
 class CommentApiController extends ApiController {
 //------------------------------------------------------------------
     public function __construct() {
@@ -37,17 +38,20 @@ class CommentApiController extends ApiController {
         $comment_id = $params[':ID'];  
         $comment = $this->model->getComment($comment_id);
 
-        $esAdmin = $this->controller->getUserBySession();
-
-        if ($comment) {
-            $this->model->deleteComment($comment_id);
-            $this->view->response("Comentario id=$comment_id eliminado con éxito", 200);
-        }
-        else 
-            $this->view->response("Comment id=$comment_id not found", 404);
+        if($this->controller->checkloggedIn()&& $_SESSION['permisoDeAdmin'] == 1){
+            if ($comment) {
+                $this->model->deleteComment($comment_id);
+                $this->view->response("Comentario id=$comment_id eliminado con exito", 200);
+            }
+            else 
+                $this->view->response("Comentario id=$comment_id no encontrado", 404);
+        }       
+        else
+            $this->view->response(" Usted no tiene permisos para realizar esta operacion", 401);
+        
     }
 //--------------------------------------------------------------
-    public function addComment() {
+    public function addComment() {//FUNCIONA BIEN.
         //AGREGA UN COMENTARIO
         $body = $this->getData();   
 
@@ -56,13 +60,17 @@ class CommentApiController extends ApiController {
         $id_producto = $body->id_producto;
         $calificacion = ($body->calificacion);
 
-        $id_Comment = $this->model->addComment($id_usuario, $comentario, $id_producto, $calificacion);
-        $comment = $this->model->getComment($id_Comment);
-
-        if ($comment)
-            $this->view->response($comment, 200);
-        else
-            $this->view->response("Error al guardar", 500);
+        if($this->controller->checkloggedIn()){
+            //verifica si es un usuario logueado. si no es asi no puede comentar.
+            $id_Comment = $this->model->addComment($id_usuario, $comentario, $id_producto, $calificacion);
+            $comment = $this->model->getComment($id_Comment);
+            if ($comment)
+                $this->view->response($comment, 200);
+            //else
+                //$this->view->response("Error al guardar", 500);
+        }else
+            $this->view->response("logue su usario para poder comentar",401);
+        
     }
 //----------------------------------------------------------------
     public function updateComment($params = []) {
@@ -83,4 +91,9 @@ class CommentApiController extends ApiController {
             $this->view->response("comment id=$comment_id not found", 404);
     }
 
+    function message($error){
+        //RENDERIZA UN MENSAJE.
+        $this->smarty->assign('error',$error);
+        $this->smarty->display('templates/message.tpl');
+    }
 }
